Refactor createTestAirfield to airfieldService
