+++ backend/src/app.ts
@@ -1,6 +1,8 @@
 import app from './index';
 import { connectToDatabase } from './util/db';
 import { PORT } from './util/config';
+import airfieldService from './services/airfieldService';
 
-connectToDatabase();
+connectToDatabase().then(airfieldService.createTestAirfield);
+
 app.listen(PORT, () => console.log(`Server is running on port: ${PORT}`));
