+++ backend/src/util/db.ts
@@ -1,22 +1,10 @@
 import { Sequelize } from 'sequelize';
 import { newDb } from 'pg-mem';
 import { DATABASE_URL } from './config';
-// eslint-disable-next-line import/no-cycle
-import { Airfield } from '../models';
 
 if (process.env.NODE_ENV === 'test') {
   console.log('creating an in-memory test database');
 }
-
-const createTestAirfield = async () => {
-  // TODO: Remove this when we have a proper admin interface for creating airfields
-  await Airfield.upsert({
-    id: 1,
-    name: 'Test Airfield',
-    maxConcurrentFlights: 3,
-    eventGranularityMinutes: 20,
-  });
-};
 
 const sequelize = process.env.NODE_ENV === 'test'
   ? new Sequelize({ dialect: 'postgres', dialectModule: newDb().adapters.createPg(), logging: false })
@@ -27,7 +15,6 @@ const connectToDatabase = async () => {
     await sequelize.sync({ alter: true });
     await sequelize.authenticate();
     console.log('database connected');
-    createTestAirfield();
   } catch (err) {
     console.log('connecting database failed. error:', err);
     return process.exit(1);
@@ -36,4 +23,4 @@ const connectToDatabase = async () => {
   return null;
 };
 
-export { connectToDatabase, sequelize, createTestAirfield };
+export { connectToDatabase, sequelize };
